// src/app/modules/api/availability/models/availability.ts
/* tslint:disable */
/* eslint-disable */
export interface Availability {
  arrivalDate?: string;
  available?: string;
  capacity?: number;
  clime?: boolean;
  date?: string;
  denomination?: string;
  price?: number;
  transport?: string;
  transportType?: string;
  travelId?: string;
  travelType?: string;
  via?: string;
}